// src/models/ConfigOTelInterface.ts
export interface ConfigOTelInterface {
  SERVICE_ID: string;
  VERSION: string;
  OPENTELEMETRY_COLLECTOR_HTTP_TRACES?: string;
  OPENTELEMETRY_COLLECTOR_HTTP_METRICS?: string;
  OPENTELEMETRY_COLLECTOR_HTTP_LOGS?: string;
  OPENTELEMETRY_COLLECTOR_EXPORT_LOGS_INTERVAL_SECONDS?: number;
  OPENTELEMETRY_COLLECTOR_EXPORT_METRICS_INTERVAL_SECONDS?: number;
  OPENTELEMETRY_COLLECTOR_AWS?: boolean;
  OPENTELEMETRY_COLLECT_AUTHORIZATION_HEADER?: string;
}